Move packages service to v2
