+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v2/IndexPackagesService.scala
@@ -3,15 +3,11 @@
 
 package com.daml.ledger.participant.state.index.v2
 
-import java.time.Instant
-
 import com.digitalasset.daml.lf.data.Ref.PackageId
 import com.digitalasset.daml_lf.DamlLf.Archive
 import com.digitalasset.daml.lf.lfpackage.Ast.Package
 
 import scala.concurrent.Future
-
-case class PackageDetails(size: Long, knownSince: Instant, sourceDescription: String)
 
 /**
   * Serves as a backend to implement
