+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v2/package.scala
@@ -86,4 +86,7 @@ package object v2 {
       workflowId: WorkflowId)
 
   final case class LedgerConfiguration(minTTL: Duration, maxTTL: Duration)
+
+  /** DAML-LF package details */
+  final case class PackageDetails(size: Long, knownSince: Instant, sourceDescription: String)
 }
