+++ ledger/participant-state/src/main/scala/com/daml/ledger/participant/state/v1/WriteService.scala
@@ -26,7 +26,7 @@ import com.digitalasset.daml_lf.DamlLf.Archive
   * - allocating a new party using [[WriteService!.allocateParty]]
   *
   */
-trait WriteService extends PackageWriteService {
+trait WriteService {
 
   /** Submit a transaction for acceptance to the ledger.
     *
