+++ ledger/participant-state/src/main/scala/com/daml/ledger/participant/state/v2/WriteService.scala
@@ -24,7 +24,7 @@ import java.util.concurrent.CompletionStage
   * - allocating a new party using [[WriteService!.allocateParty]]
   *
   */
-trait WriteService {
+trait WriteService extends WritePackagesService {
 
   /** Submit a transaction for acceptance to the ledger.
     *
